refactor InvSearchModule and InvUnloadModule to utilize Utils for block and inventory operations
